Restructured src folders for general purpose device adapter development
