+++ src/DeviceAdapters/Makefile.am
@@ -5,4 +5,4 @@ ACLOCAL_AMFLAGS = -I ../m4
 # Please keep these ASCII-lexically sorted (pass through sort(1)).
 SUBDIRS = \
 	DemoCamera \
-	RaspberryPi
+	RPiGPIO
