+++ src/DeviceAdapters/RPiGPIO/Makefile.am
@@ -1,0 +1,6 @@
+AM_CXXFLAGS = $(MMDEVAPI_CXXFLAGS)
+deviceadapter_LTLIBRARIES = libmmgr_dal_RPiGPIO.la
+libmmgr_dal_RPiGPIO_la_SOURCES = RPiGPIO.cpp RPiGPIO.h gpio.c gpio.h \
+   ../../MMDevice/MMDevice.h ../../MMDevice/DeviceBase.h
+libmmgr_dal_RPiGPIO_la_LIBADD = $(MMDEVAPI_LIBADD)
+libmmgr_dal_RPiGPIO_la_LDFLAGS = $(MMDEVAPI_LDFLAGS)
