+++ src/README.md
@@ -1,0 +1,5 @@
+# src
+
+These folders contain configuration files and source code for
+integration with Micro-Manager. The contents of this directory may be
+copied directly into the official Micro-Manager repository.
